requirements.txt
Flask
Flask-SQLAlchemy
pymysql
bcrypt
python-dotenv
mysql-connector-python
gunicorn
